Add a Runtime tracer this tracker vil track all method in scope with annotation of Tracker(name="<the name you give it>")
This wil make it easy to find the the funktions that is used on runtime.

it is logging the the data to the cmd and store it in a Feature_Logger.log.
this Feature_Logger.log need to be delete if a ned log is need, else it wil just add data to it